soft ioc function correctly
